+++ lib/souls/generate/rspec_factory.rb
@@ -4,6 +4,7 @@ module Souls
       ## Generate Rspec Factory
       def rspec_factory_head class_name: "souls"
         file_path = "./spec/factories/#{class_name.pluralize}.rb"
+        return "RspecFactory already exist! #{file_path}" if File.exist? file_path
         File.open(file_path, "w") do |f|
           f.write <<~EOS
             FactoryBot.define do
@@ -52,7 +53,6 @@ module Souls
 
       def rspec_factory class_name: "souls"
         file_path = "./spec/factories/#{class_name.pluralize}.rb"
-        return ["Factory aleady exist! #{file_path}"] if File.exist? file_path
         singularized_class_name = class_name.singularize
         rspec_factory_head class_name: singularized_class_name
         rspec_factory_params class_name: singularized_class_name
